Rename AccountCreateLegacy to AccountCreate.
