+++ assets/js/modules/tagmanager/components/common/index.js
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-export { default as AccountCreateLegacy } from './AccountCreateLegacy';
+export { default as AccountCreate } from './AccountCreate';
 export { default as AccountSelect } from './AccountSelect';
 export { default as AMPContainerSelect } from './AMPContainerSelect';
 export { default as ErrorNotice } from './ErrorNotice';
